[#22] Remove deprecated SSC types for cardano-chain
