+++ src/Cardano/Chain/Ssc.hs
@@ -4,26 +4,36 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module Cardano.Chain.Ssc
-       ( module X
-       , SscPayload (..)
-       , SscProof (..)
+       ( SscPayload (..)
        , dropSscPayload
-       , dropSscProof
-       ) where
 
-import           Cardano.Chain.Ssc.Commitment as X
-import           Cardano.Chain.Ssc.CommitmentsMap as X
-import           Cardano.Chain.Ssc.Opening as X
-import           Cardano.Chain.Ssc.OpeningsMap as X
-import           Cardano.Chain.Ssc.SharesMap as X
-import           Cardano.Chain.Ssc.VssCertificate as X
-import           Cardano.Chain.Ssc.VssCertificatesMap as X
+       , SscProof (..)
+       , dropSscProof
+
+       , dropCommitmentsMap
+       , dropSignedCommitment
+       , dropCommitment
+
+       , dropOpeningsMap
+
+       , dropSharesMap
+       , dropInnerSharesMap
+
+       , dropVssCertificatesMap
+       , dropVssCertificate
+       ) where
 
 import           Cardano.Prelude
 
 import           Cardano.Binary.Class (Bi (..), DecoderError (..), Dropper,
-                     decodeListLenCanonical, dropBytes, encodeListLen,
-                     matchSize)
+                     decodeListLenCanonical, dropBytes, dropList, dropMap,
+                     dropSet, dropTriple, dropWord64, encodeListLen,
+                     enforceSize, matchSize)
+
+
+--------------------------------------------------------------------------------
+-- SscPayload
+--------------------------------------------------------------------------------
 
 data SscPayload =
   SscPayload
@@ -59,6 +69,11 @@ dropSscPayload = do
       dropVssCertificatesMap
     t -> cborError $ DecoderErrorUnknownTag "SscPayload" t
 
+
+--------------------------------------------------------------------------------
+-- SscProof
+--------------------------------------------------------------------------------
+
 data SscProof =
   SscProof
   deriving (Eq, Show, Generic, NFData)
@@ -91,3 +106,72 @@ dropSscProof = do
       matchSize "CertificatesProof" 2 actualLen
       dropBytes
     t -> cborError $ DecoderErrorUnknownTag "SscProof" t
+
+
+--------------------------------------------------------------------------------
+-- CommitmentsMap
+--------------------------------------------------------------------------------
+
+dropCommitmentsMap :: Dropper s
+dropCommitmentsMap = dropSet dropSignedCommitment
+
+dropSignedCommitment :: Dropper s
+dropSignedCommitment = dropTriple dropBytes dropCommitment dropBytes
+
+dropCommitment :: Dropper s
+dropCommitment = do
+  enforceSize "Commitment" 2
+  -- Map (AsBinary VssPublicKey) (NonEmpty (AsBinary EncShare))
+  dropMap dropBytes (dropList dropBytes)
+  dropSecretProof
+
+dropSecretProof :: Dropper s
+dropSecretProof = do
+  enforceSize "SecretProof" 4
+  -- Scrape.ExtraGen
+  dropBytes
+  -- Scrape.Proof
+  dropBytes
+  -- Scrape.ParallelProofs
+  dropBytes
+  -- [Scrape.Commitment]
+  dropList dropBytes
+
+
+--------------------------------------------------------------------------------
+-- OpeningsMap
+--------------------------------------------------------------------------------
+
+dropOpeningsMap :: Dropper s
+dropOpeningsMap = dropMap dropBytes dropBytes
+
+
+--------------------------------------------------------------------------------
+-- SharesMap
+--------------------------------------------------------------------------------
+
+dropSharesMap :: Dropper s
+dropSharesMap = dropMap dropBytes dropInnerSharesMap
+
+dropInnerSharesMap :: Dropper s
+dropInnerSharesMap = dropMap dropBytes (dropList dropBytes)
+
+
+--------------------------------------------------------------------------------
+-- VssCertificatesMap
+--------------------------------------------------------------------------------
+
+dropVssCertificatesMap :: Dropper s
+dropVssCertificatesMap = dropSet dropVssCertificate
+
+dropVssCertificate :: Dropper s
+dropVssCertificate = do
+  enforceSize "VssCertificate" 4
+  -- AsBinary VssPublicKey
+  dropBytes
+  -- EpochIndex
+  dropWord64
+  -- Signature (AsBinary VssPublicKey, EpochIndex)
+  dropBytes
+  -- PublicKey
+  dropBytes
