+++ test/Test/Cardano/Chain/Ssc/Bi.hs
@@ -8,7 +8,6 @@ module Test.Cardano.Chain.Ssc.Bi
 import           Cardano.Prelude
 import           Test.Cardano.Prelude
 
-import qualified Data.Map.Strict as Map
 import           Hedgehog (Property)
 import qualified Hedgehog as H
 
@@ -19,45 +18,22 @@ import           Cardano.Chain.Ssc (SscPayload (..), SscProof (..),
                      dropSscPayload, dropSscProof, dropVssCertificate,
                      dropVssCertificatesMap)
 
-import           Test.Cardano.Binary.Helpers.GoldenRoundTrip (goldenTestBi,
-                     legacyGoldenDecode, roundTripsBiBuildable,
-                     roundTripsBiShow)
-import           Test.Cardano.Chain.Common.Example (exampleStakeholderId)
-import           Test.Cardano.Chain.Ssc.Example (exampleCommitment,
-                     exampleCommitmentSignature, exampleCommitmentsMap,
-                     exampleInnerSharesMap, exampleOpening, exampleOpeningsMap,
-                     exampleSignedCommitment, exampleVssCertificate,
-                     exampleVssCertificatesMap)
-import           Test.Cardano.Chain.Ssc.Gen (genCommitment,
-                     genCommitmentSignature, genCommitmentsMap,
-                     genInnerSharesMap, genOpening, genOpeningsMap,
-                     genSharesMap, genSignedCommitment, genVssCertificate,
-                     genVssCertificatesMap)
-import           Test.Cardano.Crypto.Gen (feedPM)
+import           Test.Cardano.Binary.Helpers.GoldenRoundTrip
+                     (legacyGoldenDecode, roundTripsBiShow)
 
 
 --------------------------------------------------------------------------------
 -- Commitment
 --------------------------------------------------------------------------------
 
-golden_Commitment :: Property
-golden_Commitment = goldenTestBi exampleCommitment "test/golden/Commitment"
-
 golden_legacy_Commitment :: Property
 golden_legacy_Commitment =
   legacyGoldenDecode "Commitment" dropCommitment "test/golden/Commitment"
-
-roundTripCommitment :: Property
-roundTripCommitment = eachOf 10 genCommitment roundTripsBiShow
 
 
 --------------------------------------------------------------------------------
 -- CommitmentsMap
 --------------------------------------------------------------------------------
-
-golden_CommitmentsMap :: Property
-golden_CommitmentsMap =
-  goldenTestBi exampleCommitmentsMap "test/golden/CommitmentsMap"
 
 golden_legacy_CommitmentsMap :: Property
 golden_legacy_CommitmentsMap = legacyGoldenDecode
@@ -65,30 +41,10 @@ golden_legacy_CommitmentsMap = legacyGoldenDecode
   dropCommitmentsMap
   "test/golden/CommitmentsMap"
 
-roundTripCommitmentsMap :: Property
-roundTripCommitmentsMap = eachOf 10 (feedPM genCommitmentsMap) roundTripsBiShow
-
-
---------------------------------------------------------------------------------
--- CommitmentsSignature
---------------------------------------------------------------------------------
-
-golden_CommitmentSignature :: Property
-golden_CommitmentSignature =
-  goldenTestBi exampleCommitmentSignature "test/golden/CommitmentSignature"
-
-roundTripCommitmentSignature :: Property
-roundTripCommitmentSignature =
-  eachOf 10 (feedPM genCommitmentSignature) roundTripsBiBuildable
-
 
 --------------------------------------------------------------------------------
 -- InnerSharesMap
 --------------------------------------------------------------------------------
-
-golden_InnerSharesMap :: Property
-golden_InnerSharesMap = goldenTestBi iSm "test/golden/InnerSharesMap"
-  where iSm = exampleInnerSharesMap 3 1
 
 golden_legacy_InnerSharesMap :: Property
 golden_legacy_InnerSharesMap = legacyGoldenDecode
@@ -96,31 +52,19 @@ golden_legacy_InnerSharesMap = legacyGoldenDecode
   dropInnerSharesMap
   "test/golden/InnerSharesMap"
 
-roundTripInnerSharesMap :: Property
-roundTripInnerSharesMap = eachOf 50 genInnerSharesMap roundTripsBiShow
-
 
 --------------------------------------------------------------------------------
 -- Opening
 --------------------------------------------------------------------------------
 
-golden_Opening :: Property
-golden_Opening = goldenTestBi exampleOpening "test/golden/Opening"
-
 golden_legacy_Opening :: Property
 golden_legacy_Opening =
   legacyGoldenDecode "Opening" dropBytes "test/golden/Opening"
-
-roundTripOpening :: Property
-roundTripOpening = eachOf 10 genOpening roundTripsBiBuildable
 
 
 --------------------------------------------------------------------------------
 -- OpeningsMap
 --------------------------------------------------------------------------------
-
-golden_OpeningsMap :: Property
-golden_OpeningsMap = goldenTestBi exampleOpeningsMap "test/golden/OpeningsMap"
 
 golden_legacy_OpeningsMap :: Property
 golden_legacy_OpeningsMap = legacyGoldenDecode
@@ -128,17 +72,10 @@ golden_legacy_OpeningsMap = legacyGoldenDecode
   dropOpeningsMap
   "test/golden/OpeningsMap"
 
-roundTripOpeningsMap :: Property
-roundTripOpeningsMap = eachOf 10 genOpeningsMap roundTripsBiShow
-
 
 --------------------------------------------------------------------------------
 -- SignedCommitment
 --------------------------------------------------------------------------------
-
-golden_SignedCommitment :: Property
-golden_SignedCommitment =
-  goldenTestBi exampleSignedCommitment "test/golden/SignedCommitment"
 
 golden_legacy_SignedCommitment :: Property
 golden_legacy_SignedCommitment = legacyGoldenDecode
@@ -146,27 +83,16 @@ golden_legacy_SignedCommitment = legacyGoldenDecode
   dropSignedCommitment
   "test/golden/SignedCommitment"
 
-roundTripSignedCommitment :: Property
-roundTripSignedCommitment =
-  eachOf 10 (feedPM genSignedCommitment) roundTripsBiShow
-
 
 --------------------------------------------------------------------------------
 -- SharesMap
 --------------------------------------------------------------------------------
-
-golden_SharesMap :: Property
-golden_SharesMap = goldenTestBi sM "test/golden/SharesMap"
-  where sM = Map.fromList $ [(exampleStakeholderId, exampleInnerSharesMap 3 1)]
 
 golden_legacy_SharesMap :: Property
 golden_legacy_SharesMap = legacyGoldenDecode
   "SharesMap"
   dropSharesMap
   "test/golden/SharesMap"
-
-roundTripSharesMap :: Property
-roundTripSharesMap = eachOf 10 genSharesMap roundTripsBiShow
 
 
 --------------------------------------------------------------------------------
@@ -237,19 +163,11 @@ roundTripSscProof = eachOf 1 (pure SscProof) roundTripsBiShow
 -- VssCertificate
 --------------------------------------------------------------------------------
 
-golden_VssCertificate :: Property
-golden_VssCertificate =
-  goldenTestBi exampleVssCertificate "test/golden/VssCertificate"
-
 golden_legacy_VssCertificate :: Property
 golden_legacy_VssCertificate = legacyGoldenDecode
   "VssCertificate"
   dropVssCertificate
   "test/golden/VssCertificate"
-
-roundTripVssCertificate :: Property
-roundTripVssCertificate =
-  eachOf 10 (feedPM genVssCertificate) roundTripsBiBuildable
 
 
 --------------------------------------------------------------------------------
@@ -267,19 +185,11 @@ golden_legacy_VssCertificatesHash = legacyGoldenDecode
 -- VssCertificatesMap
 --------------------------------------------------------------------------------
 
-golden_VssCertificatesMap :: Property
-golden_VssCertificatesMap =
-  goldenTestBi (exampleVssCertificatesMap 10 4) "test/golden/VssCertificatesMap"
-
 golden_legacy_VssCertificatesMap :: Property
 golden_legacy_VssCertificatesMap = legacyGoldenDecode
   "VssCertificatesMap"
   dropVssCertificatesMap
   "test/golden/VssCertificatesMap"
-
-roundTripVssCertificatesMap :: Property
-roundTripVssCertificatesMap =
-  eachOf 10 (feedPM genVssCertificatesMap) roundTripsBiShow
 
 
 --------------------------------------------------------------------------------
